+++ test/copy_directory_test.go
@@ -2,6 +2,7 @@
 package test
 
 import (
+	"os"
 	"path/filepath"
 	"testing"
 
@@ -36,4 +37,14 @@ func TestCopyDirectory(t *testing.T) {
 		asserts.FileExists(t, dstDir, "one.txt")
 		asserts.FileHasContent(t, dstDir, ".git/HEAD", "ref: refs/heads/initial\n")
 	})
+}
+
+// createFile creates a file with the given filename in the given directory.
+func createFile(t *testing.T, dir, filename string) {
+	t.Helper()
+	filePath := filepath.Join(dir, filename)
+	err := os.MkdirAll(filepath.Dir(filePath), 0o744)
+	assert.NoError(t, err)
+	err = os.WriteFile(filePath, []byte(filename+" content"), 0o500)
+	assert.NoError(t, err)
 }
